Update wal transformer processor with builder
